Fix bug with login component, update graphql queries, start adding non-editor pages
