+++ codegen.yml
@@ -1,5 +1,5 @@
 overwrite: true
-schema: "http://localhost:4002"
+schema: "http://localhost:4001"
 documents: "src/graphql/*"
 generates:
   src/generated/queries.ts:
