--- src/graphql/getUpdates.gql
@@ -1,5 +1,0 @@
-query GetUpdates {
-    updates {
-        title, contents, date
-    }
-}
