+++ src/graphql/saveProject.gql
@@ -1,6 +1,5 @@
 mutation SaveProject($id: String!, $projectJson: String!) {
   editProject(id: $id, projectJson: $projectJson) {
-    success,
-    nModified
+    id
   }
 }
