--- src/pages/Editor.css
@@ -1,32 +1,0 @@
-.gridContainer {
-  display: flex;
-  height: calc(100vh - 125px);
-  padding: 20px;
-}
-
-.editorColumn {
-  flex: 1;
-}
-
-.toolsColumn {
-  display: flex;
-  flex-direction: column;
-  max-width: 40rem;
-  width: 40vw;
-}
-
-.toolsColumn > * {
-  flex: 1;
-}
-
-.codeEditor {
-  height: 100%;
-}
-
-.gridContainer .MuiPaper-root {
-  margin: 0.5rem;
-}
-
-.saveModalCodeContainer {
-  overflow-x: scroll;
-}
