+++ src/pages/dashboard.tsx
@@ -1,0 +1,63 @@
+import { Typography } from "@material-ui/core";
+import PageLayout from "../components/PageLayout";
+
+export default function Dashboard() {
+    return (
+        <PageLayout>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+            <Typography variant="h1">Hello world!</Typography>
+
+        </PageLayout>
+    );
+}
