+++ src/pages/project/[id].tsx
@@ -1,0 +1,5 @@
+export default function Project() {
+    return (
+        <div>Todo</div>
+    );
+}
