+++ src/utils/apolloClient.ts
@@ -1,6 +1,7 @@
 import { ApolloClient, createHttpLink, InMemoryCache, NormalizedCacheObject } from "@apollo/client";
 import { setContext } from '@apollo/client/link/context';
 import { useMemo } from "react";
+import firebase from "./firebase";
 
 let apolloClient: ApolloClient<NormalizedCacheObject> | undefined;
 
@@ -9,9 +10,8 @@ const httpLink = createHttpLink({
 });
 
 const authLink = setContext(async (_, { headers }) => {
-    // TODO
-    return {};
-    /*const token = await fetchToken();
+    const token = await firebase.auth().currentUser?.getIdToken();
+    console.log(token);
     if(token === undefined) return {};
 
     return {
@@ -19,7 +19,7 @@ const authLink = setContext(async (_, { headers }) => {
             ...headers,
             authorization: token
         }
-    }*/
+    }
 });
 
 const createApolloClient = () => new ApolloClient({
